HotelManagement(V1)
This version is just the first deploy of system, i made a primary menu, where you can choose one of two options, "Novo Cadastro" for realise a new register, or "Já tenho cadastro", for you made login with your credentials. When you made a new register, you e-mail and password will saved in the data base, when the register will made, you just need to click on "Ja tenho cadastro" and be happy!